# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores:
    get:
      operationId: listStores
      tags: [stores]
      summary: List all stores
      responses:
        "200":
          description: List of stores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Store"
    post:
      operationId: createStore
      tags: [stores]
      summary: Create a new store
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreInput"
      responses:
        "201":
          description: Created store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
  /stores/{id}:
    get:
      operationId: getStore
      tags: [stores]
      summary: Get a store by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Store details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateStore
      tags: [stores]
      summary: Update a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StoreUpdate"
      responses:
        "200":
          description: Updated store
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Store"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    delete:
      operationId: deleteStore
      tags: [stores]
      summary: Delete a store
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # Chemicals
...
  /inventory/{storeId}:
    get:
      operationId: getStoreInventory
      tags: [inventory]
      summary: Get inventory counts for a specific store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory for the store
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InventoryEntry"
  /inventory/{storeId}/{chemicalId}:
    put:
      operationId: upsertInventoryCount
      tags: [inventory]
      summary: Set (create or update) inventory count for a store+chemical
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: chemicalId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryCountInput"
      responses:
        "200":
          description: Updated inventory entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryEntry"
    delete:
      operationId: deleteInventoryCount
      tags: [inventory]
      summary: Remove inventory count for a store+chemical pair
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: chemicalId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Dashboard summary
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Stores
    Store:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    StoreInput:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        notes:
          type: string
      required:
        - name
    StoreUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        phone:
          type: string
        notes:
          type: string

    # Chemicals
...
    InventoryEntry:
      type: object
      properties:
        storeId:
          type: integer
        chemicalId:
          type: integer
        quantity:
          type: number
        storeName:
          type: string
        chemicalName:
          type: string
        unit:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - storeId
        - chemicalId
        - quantity
        - storeName
        - chemicalName
        - unit
        - updatedAt
    InventoryCountInput:
      type: object
      properties:
        quantity:
          type: number
      required:
        - quantity
...