# lib/api-zod/src/generated/types/inventoryEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryEntry {
  storeId: number;
  chemicalId: number;
  quantity: number;
  storeName: string;
  chemicalName: string;
  unit: string;
  updatedAt: Date;
}